This commit includes the html, css, and js files that makeup the app.
Note:

For later updates, there is a small bug that needs to be fixed upon reloading the page. The previous tasks are restored but the style of a 'checked' box is removed after reloading and subsequently, the checkbox is removed as well upon a third reload. However, the tasks are restored properly.